refactor : 페이지네이션 컴포넌트 재사용을 위해 페이지 변경 의존성 주입
